finished adding eslint semi colon and some hooks for pages, still missing hooks for components with extended functionality
