+++ src/components/404/Star.tsx
@@ -1,5 +1,3 @@
-import React from 'react'
-
 export const Star = () => {
-    return (<div className='star'></div>)
-}
+    return (<div className='star'></div>);
+};
