+++ src/components/Form/index.tsx
@@ -1,5 +1,5 @@
 import { FormikHandlers, } from 'formik';
-import React, { ReactNode } from 'react'
+import { ReactNode } from 'react';
 import terminalFormStyles from './Form.module.css';
 import newsletterStyles from '../Newsletter/Newsletter.module.css';
 import layoutStyles from '../Layout/Layout.module.css';
@@ -14,7 +14,7 @@ type FormProps = {
 export const Form = ({ children, handleSubmit, terminal = false, newsletter = false}: FormProps) => {
 	return (
 		<form 
-			onSubmit={(e) => {e.preventDefault(); if(handleSubmit) handleSubmit(e)}}
+			onSubmit={(e) => {e.preventDefault(); if(handleSubmit) handleSubmit(e);}}
 			autoComplete='off'
 			className={`
 				${terminal ? terminalFormStyles.terminalForm : ''}
@@ -24,5 +24,5 @@ export const Form = ({ children, handleSubmit, terminal = false, newsletter = fa
 		>
 			{children}
 		</form>
-	)
-}
+	);
+};
