+++ src/components/Index/text.ts
@@ -1,7 +1,14 @@
 export const texts = {
-    heading: '',
+    heading: {
+        firstPart: 'Artesanía convertida en ',
+        highlight: 'tecnología',
+        secondPart: ' que cosechan tus ideas',
+    },
     subheading: 'Fundados en 2022, Jardín Binario busca ayudar a empresas en su inversión tecnológica, creando herramientas para aumentar su máximo potencial de mercado.',
-    heading2: '',
+    heading2: {
+        firstPart: 'Un equipo que busca ayudarte a ',
+        highlight: 'analizar, sembrar y cosechar',
+        secondPart: ' tus ideas',
+    },
     subheading2: 'Nuestros ingenieros y diseñadores, han establecido procesos rigurosos y comunicativos, donde comenzamos por un diseño de marca para el cliente, luego por maquetaciones y finalmente la implementación de la solución con todos sus subprocesos técnicos. Y claro, todo esto de la mano del cliente, siempre anuentes y dispuestos a cualquier sugerencia.',
-    heading3: 'Nuestros servicios:',
 };
