+++ src/components/Newsletter/Newsletter.module.css
@@ -14,7 +14,7 @@
 }
 
 .copy {
-    @apply text-xs mt-3 md:mt-0 text-left pb-0 mb-0;
+    @apply text-xs mt-3 text-left pb-0 mb-0;
     text-align: left!important;
 }
