+++ src/pages/404.tsx
@@ -1,10 +1,11 @@
 import { useRouter } from 'next/router';
-import React, { useEffect } from 'react';
+import { useEffect } from 'react';
 import { Sky } from '../components/404/Sky';
 import { Layout } from '../components/Layout';
 
 export default function Custom404() {
 	const router = useRouter();
+
 	useEffect(() => {
 		setTimeout(() => router.push('/'), 6500);
 	}, [router]);
@@ -20,5 +21,5 @@ export default function Custom404() {
 				</div>
 			</Layout>
 		</>
-	)
+	);
 }
