+++ src/pages/_document.tsx
@@ -1,5 +1,5 @@
 
-import { Html, Head, Main, NextScript } from 'next/document'
+import { Html, Head, Main, NextScript } from 'next/document';
 
 export default function Document() {
 	return (
@@ -15,5 +15,5 @@ export default function Document() {
 				<NextScript />
 			</body>
 		</Html>
-	)
+	);
 }
