+++ src/pages/index.tsx
@@ -1,25 +1,21 @@
+import { InferGetServerSidePropsType } from 'next';
+
 import { Layout } from '../components/Layout';
 import { TerminalHeader } from '../components/Terminal/TerminalHeader';
-import { RefObject, useEffect, useRef, useState } from 'react';
-import { useFormik } from 'formik';
-import { useMutation } from '@apollo/client';
-import * as Yup from 'yup';
-import { InferGetServerSidePropsType } from 'next';
 import { Footer } from '../components/Footer';
-
 import { texts } from '../components/Index/text';
 import indexStyles from '../components/Index/Index.module.css';
 import { PhotoComposition } from '../components/PhotoComposition';
 import { Services } from '../components/Services';
 import { TicketForm } from '../components/TicketForm';
 import { querys } from '../gql/querys';
-import { NewsletterValues, NewTicketValues } from '../types/sharedTypes';
 import { Newsletter } from '../components/Newsletter';
 import { CustomSwiper } from '../components/Swiper';
 import { createUnauthorizedApolloClient } from '../apollo/AuthClient';
-import { useRouter } from 'next/router';
+import { useIndex } from '../hooks/useIndex';
+import { HeadingBlock } from '../components/Index/HeadingBlock';
 
-export const getServerSideProps = async (context: any) => {
+export const getServerSideProps = async () => {
 	const client = createUnauthorizedApolloClient();
 
 	const { data: { getRecentEntries } } = await client.query({
@@ -33,7 +29,7 @@ export const getServerSideProps = async (context: any) => {
 	if (!getRecentEntries || !getMostViewedEntries) {
 		return {
 			notFound: true,
-		}
+		};
 	}
 
 	return {
@@ -42,149 +38,29 @@ export const getServerSideProps = async (context: any) => {
 			mostViewedEntries: getMostViewedEntries
 		}
 	};
-}
-
-function timeout(delay: number) {
-	return new Promise(res => setTimeout(res, delay));
-}
+};
 
 export default function IndexPage({ recentEntries, mostViewedEntries }: InferGetServerSidePropsType<typeof getServerSideProps>) {
-	const [newTicket] = useMutation(querys.NEW_TICKET);
-	const [newNewsletter] = useMutation(querys.NEW_NEWSLETTER);
-
-	const router = useRouter();
-	// TODO maybe get a hook  going?
-	const sharedState = (initialState: boolean) => ({
-		contactForm: initialState,
-		newsletterForm: initialState
-	});
-
-	const [disableButton, setDisableButton] = useState<Record<string, boolean>>(sharedState(true));
-	const [submitted, setSubmitted] = useState<Record<string, boolean>>(sharedState(false));
-
-	const [message, setMessage] = useState<string>('');
-
-	const initialValuesContactForm: NewTicketValues = {
-		companyName: '',
-		email: '',
-		description: '',
-		service: [],
-		phoneNumber: '',
-	};
-
-	const initialValuesNewsletter: NewsletterValues = {
-		email: '',
-	};
-
-	const phoneRegExp = /^[0-9]{4}[-][0-9]{4}$/;
-
-	const formikContactForm = useFormik({
-		initialValues: initialValuesContactForm,
-		validationSchema: Yup.object({
-			companyName: Yup.string().required('Un nombre de compañia es requerido'),
-			email: Yup.string().email().required('Un correo electronico es requerido'),
-			service: Yup.array().of(Yup.string()).min(1),
-			description: Yup.string().required('Una descripcion es requerida'),
-			phoneNumber: Yup.string().matches(phoneRegExp, 'El formato del numero es invalido').required('Un numero telefonico es requerido'),
-		}),
-		enableReinitialize: true,
-		onSubmit: async (values) => {
-			try {
-				await timeout(4000);
-				const response = await newTicket({
-					variables: {
-						ticketInput: {
-							...values,
-						}
-					}
-				});
-
-				if (response.data.newTicket) {
-					let submittedCopy = { ...submitted };
-					submittedCopy.contactForm = true;
-					setSubmitted(submittedCopy);
-					setMessage('Gracias por tu mensaje, estaremos en contacto muy pronto!')
-				}
-			} catch (err: any) {
-				// do we set error here 
-				console.error(err);
-			}
-		},
-	});
-
-	const formikNewsletter = useFormik({
-		initialValues: initialValuesNewsletter,
-		validationSchema: Yup.object({
-			email: Yup.string().email().required('Un correo electronico es requerido'),
-		}),
-		onSubmit: async (values) => {
-			try {
-				await timeout(4000);
-				const response = await newNewsletter({
-					variables: {
-						newsletterInput: {
-							...values,
-						}
-					}
-				});
-
-				if (response.data.newNewsletterEntry) {
-					let submittedCopy = { ...submitted };
-					submittedCopy.newsletterForm = true;
-					setSubmitted(submittedCopy);
-					// setMessage('Gracias por tu mensaje, estaremos en contacto muy pronto!');
-				}
-			} catch (err: any) {
-				// do we set error here 
-				console.error(err);
-			}
-		},
-	});
-
-	const refServices = useRef<HTMLDivElement>(null);
-	const refForm = useRef<HTMLDivElement>(null);
-
-	const handleClickServices = (ref: string) => {
-		const toScrollMapping: Record<string, RefObject<HTMLDivElement>> = {
-			ticket: refForm,
-			services: refServices,
-		};
-
-		const toScroll = toScrollMapping[ref];
-		toScroll?.current?.scrollIntoView({
-			behavior: 'smooth',
-		});
-	};
-
-	useEffect(() => {
-		let disabledButtonCopy = { ...disableButton };
-
-		if (formikContactForm.isValid) {
-			disabledButtonCopy.contactForm = false;
-		} else {
-			disabledButtonCopy.contactForm = true;
-		}
-
-		if (formikNewsletter.isValid) {
-			disabledButtonCopy.newsletterForm = false;
-		} else {
-			disabledButtonCopy.newsletterForm = true;
-		}
-
-		setDisableButton(disabledButtonCopy);
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [formikContactForm.isValid, formikNewsletter.isValid]);
+	const {
+		disableButton,
+		formikContactForm,
+		formikNewsletter,
+		refServices,
+		refForm,
+		handleClickServices,
+		submitted,
+		message,
+		// TODO think about a router HOC or a way to pass this easily to instanciate it just once on <App/> level maybe?
+		router
+	} = useIndex();
 
 	return (
 		<Layout index>
 			<TerminalHeader router={router} handleClickServices={handleClickServices} index header='Jardín Binario' />
 			<div className={indexStyles.index}>
-				<h1>Artesanía convertida en <span className={indexStyles.heading}>tecnología</span> que cosechan tus ideas</h1>
-				<p>{texts.subheading}</p>
+				<HeadingBlock tag='h1' heading={texts.heading} subheading={texts.subheading} />
 				<PhotoComposition />
-				<h2 className='mt-0'>Un equipo que busca ayudarte a analizar, <span className={indexStyles.heading}>sembrar y cosechar</span> tus ideas.</h2>
-				<p>{texts.subheading2}</p>
-				<h3>{texts.heading3}</h3>
+				<HeadingBlock tag='h2' heading={texts.heading2} subheading={texts.subheading2} />
 				<Services refForScroll={refServices} />
 				<TicketForm
 					handleChange={formikContactForm.handleChange}
@@ -209,8 +85,7 @@ export default function IndexPage({ recentEntries, mostViewedEntries }: InferGet
 				/>
 				<CustomSwiper router={router} recentBlogs={recentEntries} mostViewedBlogs={mostViewedEntries} />
 			</div>
-			<Footer router={router}/>
+			<Footer router={router} />
 		</Layout>
-
-	)
+	);
 }
