+++ src/pages/login.tsx
@@ -1,7 +1,3 @@
-import { useFormik } from 'formik';
-import { useEffect, useMemo, useState } from 'react';
-import { useRouter } from 'next/router';
-import * as Yup from 'yup';
 import { HelpMessage } from '../components/HelpMessage';
 import { Layout } from '../components/Layout';
 import { Terminal } from '../components/Terminal/Terminal';
@@ -9,46 +5,17 @@ import { TerminalButton } from '../components/Terminal/TerminalButton';
 import { TerminalHeader } from '../components/Terminal/TerminalHeader';
 import { TerminalInput } from '../components/Terminal/TerminalInput';
 import { Form } from '../components/Form';
-import { useAuth } from '../apollo/AuthClient';
-
-type LoginFormValues = {
-	email: string;
-	password: string;
-};
+import { LoginFormValues } from '../types/sharedTypes';
+import { useLogin } from '../hooks/useLogin';
 
 export default function LoginPage() {
-	const [disableButton, setDisableButton] = useState<boolean>(true);
-
-	const initialValues: LoginFormValues = useMemo(() => ({
-		email: '',
-		password: ''
-	}), []);
-
-	const { signIn } = useAuth();
-	const router = useRouter();
-
-	const formik = useFormik({
+	
+	const {
+		disableButton,
+		formik,
 		initialValues,
-		validationSchema: Yup.object({
-			email: Yup.string().email().required('An email is required'),
-			password: Yup.string().required('A password name is required').min(7, 'The password should at least have 7 characters'),
-		}),
-		onSubmit: async (values) => {
-			try{
-				await signIn(values);
-				router.push('/new');
-			} catch(err) {
-				console.error(err);
-			}
-		},
-	});
-
-	useEffect(() => {
-		if (formik.isValid && formik.dirty) {
-			//TODO do the fingerprint here and see how to append the fingerprint data into the request.
-			setDisableButton(false);
-		} else setDisableButton(true);
-	}, [formik]);
+		router
+	} = useLogin();
 
 	return (
 		<>
@@ -85,5 +52,5 @@ export default function LoginPage() {
 				</div>
 			</Layout>
 		</>
-	)
+	);
 }
