+++ src/tests/components/Message.test.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
 import { Message } from '../../components/Message';
