+++ src/tests/components/Tooltip.test.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
 import { Tooltip } from '../../components/Tooltip';
